Erro ao atualizar perfil de comitê em usuário com perfil de operador #349
Erro ao atualizar perfil de comitê em usuário com perfil de operador #349

+++ app/Data/Repositories/UsersCommittees.php
@@ -55,7 +55,8 @@ class UsersCommittees extends Base
 
     public function deleteUserCommittee($user_id, $committee_id)
     {
-        $row = $this->searchByUserAndCommittee($user_id, $committee_id);
-        $row->delete();
+        if ($row = $this->searchByUserAndCommittee($user_id, $committee_id)) {
+            $row->delete();
+        }
     }
 }
